moved OneVsRest from reduction subpackage to classification.  small java doc style fixes
